+++ coaster/views.py
@@ -4,7 +4,8 @@ from __future__ import absolute_import
 from functools import wraps
 import urlparse
 import re
-from flask import session as request_session, request, url_for, json, Response, redirect, abort, g
+from flask import (session as request_session, request, url_for, json, Response,
+    redirect, abort, g, current_app, render_template)
 from werkzeug.routing import BuildError
 from werkzeug.exceptions import BadRequest
 from sqlalchemy.orm.exc import NoResultFound
@@ -368,5 +369,127 @@ def load_models(*chain, **kwargs):
                     return f(kwargs=kw, **result)
                 else:
                     return f(**result)
+        return decorated_function
+    return inner
+
+
+def render_with(template):
+    """
+    Decorator to render the wrapped method with the given template (or dictionary
+    of mimetype keys to templates, where the template is a string name of a template
+    file or a callable that returns a Response). The method's return value must be
+    a dictionary and is passed to the template as parameters. Callable templates get
+    a single parameter with the method's return value. Usage::
+
+        @app.route('/myview')
+        @render_with('myview.html')
+        def myview():
+            return {'data': 'value'}
+
+        @app.route('/otherview')
+        @render_with({
+            'text/html': 'otherview.html',
+            'text/xml': 'otherview.xml'})
+        def otherview():
+            return {'data': 'value'}
+
+    When a mimetype is specified and the template is not a callable, the response is
+    returned with the same mimetype. Callable templates must return Response objects
+    to ensure the correct mimetype is set.
+
+    If the method is called outside a request context, the wrapped method's original
+    return value is returned. This is meant to facilitate testing and should not be
+    used to call the method from within another view handler as the presence of a
+    request context will trigger template rendering.
+
+    Rendering may also be suspended by calling the view handler with ``_render=False``.
+
+    render_with provides a default handler for the ``application/json``, ``text/json``
+    and ``text/x-json`` mimetypes.
+    """
+    templates = {
+        'application/json': jsonp,
+        'text/json': jsonp,
+        'text/x-json': jsonp,
+        }
+    if isinstance(template, basestring):
+        templates['*/*'] = template
+    elif isinstance(template, dict):
+        templates.update(template)
+    else:  # pragma: no cover
+        raise ValueError("Expected string or dict for template")
+
+    def inner(f):
+        @wraps(f)
+        def decorated_function(*args, **kwargs):
+            # Check if we need to bypass rendering
+            render = kwargs.pop('_render', True)
+
+            # Get the result
+            result = f(*args, **kwargs)
+
+            # Did the result include status code and headers?
+            if isinstance(result, tuple):
+                resultset = result
+                result = resultset[0]
+                if len(resultset) > 1:
+                    status_code = resultset[1]
+                else:
+                    status_code = None
+                if len(resultset) > 2:
+                    headers = resultset[2]
+                else:
+                    headers = None
+            else:
+                status_code = None
+                headers = None
+
+            # Find a matching mimetype between Accept headers and available templates
+            use_mimetype = None
+            if render:
+                try:
+                    mimetypes = [m.strip() for m in request.headers.get(
+                        'Accept', '').replace(';', ',').split(',') if '/' in m]
+                    use_mimetype = None
+                    for mimetype in mimetypes:
+                        if mimetype in templates:
+                            use_mimetype = mimetype
+                            break
+                    if use_mimetype is None:
+                        if '*/*' in templates:
+                            use_mimetype = '*/*'
+                except RuntimeError:  # Not in a request context
+                    pass
+
+            # Now render the result with the template for the mimetype
+            if use_mimetype is not None:
+                if callable(templates[use_mimetype]):
+                    rendered = templates[use_mimetype](result)
+                    if isinstance(rendered, Response):
+                        if status_code is not None:
+                            rendered.status_code = status_code
+                        if headers is not None:
+                            rendered.headers.update(headers)
+                    else:
+                        rendered = current_app.response_class(
+                            rendered,
+                            status_code=status_code,
+                            headers=headers,
+                            mimetype=use_mimetype)
+                else:
+                    if use_mimetype != '*/*':
+                        rendered = current_app.response_class(
+                            render_template(templates[use_mimetype], **result),
+                            status_code=status_code, headers=headers,
+                            mimetype=use_mimetype)
+                    else:
+                        rendered = render_template(templates[use_mimetype], **result)
+                        if status_code is not None and headers is not None:
+                            rendered = (rendered, status_code, headers)
+                        elif status_code is not None:
+                            rendered = (rendered, status_code)
+                return rendered
+            else:
+                return result
         return decorated_function
     return inner